+++ lib/src/scope.dart
@@ -221,7 +221,8 @@ class Scope {
       // Remove all children not existing in replacement anymore
       final removedChildren = oldScope.children
           .where(
-            (element) => !replacement.children.any((c) => c.key == element.key),
+            (element) =>
+                !replacement.scopeOverrides.any((c) => c.key == element.key),
           )
           .toList();
 
@@ -230,7 +231,7 @@ class Scope {
       }
 
       // Also replace child scopes
-      for (final newChildScope in replacement.children) {
+      for (final newChildScope in replacement.scopeOverrides) {
         // Get the associated oldChildScope
         final oldChildScope = oldScope.child(newChildScope.key);
 
@@ -809,15 +810,15 @@ class Scope {
   void _updateNodesInScope(Scope previous, ScopeBluePrint current) {
     // ....................
     // Estimate added nodes
-    final addedNodes =
-        current.nodes.where((c) => !previous.nodes.any((p) => c.key == p.key));
+    final addedNodes = current.nodeOverrides
+        .where((c) => !previous.nodes.any((p) => c.key == p.key));
 
     // Estimate removed nodes
-    final removedNodes =
-        previous.nodes.where((p) => !current.nodes.any((c) => c.key == p.key));
+    final removedNodes = previous.nodes
+        .where((p) => !current.nodeOverrides.any((c) => c.key == p.key));
 
     // Estimate changed nodes
-    final changedNodes = current.nodes.where(
+    final changedNodes = current.nodeOverrides.where(
       (c) => previous.nodes.any(
         (p) => c.key == p.key && c != p.bluePrint,
       ),